requirements.txt
Django~=2.0.6
python-dateutil~=2.8.0